Refactor prog commands to separate parser class
